feat: Implement Baseline Mode with new mathematical services, data models, and WPF UI components for data processing and visualization.

+++ Core/Models/Baseline/BaselineData.cs
@@ -1,20 +1,18 @@
-using System.Collections.Generic;
-
 namespace BaselineMode.WPF.Core.Models.Baseline
 {
     public class BaselineData
     {
         public int SamplingPacketNo { get; set; }
         public int SamplingNo { get; set; }
-        public double[] L1 { get; set; } = new double[16];
-        public double[] L2 { get; set; } = new double[16];
-        public double[] L6 { get; set; } = new double[16];
-        public double[] L7 { get; set; } = new double[16];
+        public float[] L1 { get; set; } = new float[16];
+        public float[] L2 { get; set; } = new float[16];
+        public float[] L6 { get; set; } = new float[16];
+        public float[] L7 { get; set; } = new float[16];
 
         // Voltage
-        public double[] L1_Voltage { get; set; } = new double[16];
-        public double[] L2_Voltage { get; set; } = new double[16];
-        public double[] L6_Voltage { get; set; } = new double[16];
-        public double[] L7_Voltage { get; set; } = new double[16];
+        public float[] L1_Voltage { get; set; } = new float[16];
+        public float[] L2_Voltage { get; set; } = new float[16];
+        public float[] L6_Voltage { get; set; } = new float[16];
+        public float[] L7_Voltage { get; set; } = new float[16];
     }
 }
